# openapi/management-open-api.yaml
openapi: 3.0.3
info:
  title: iceberg-catalog
  description: |
    Implementation of the Iceberg REST Catalog server.
  license:
    name: Apache-2.0
  version: 0.4.0
paths:
  /management/v1/project:
    get:
      tags:
      - management
      summary: List all projects the requesting user has access to
      operationId: list_projects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListProjectsResponse'
    post:
      tags:
      - management
      summary: Create a new project
      operationId: create_project
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
        required: true
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
  /management/v1/project/{project_id}:
    get:
      tags:
      - management
      summary: Get a Project by ID
      operationId: get_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetProjectResponse'
    delete:
      tags:
      - management
      summary: Delete a project by ID
      description: No warehouses must be present in the project to delete it.
      operationId: delete_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Project deleted successfully
  /management/v1/project/{project_id}/rename:
    post:
      tags:
      - management
      summary: Rename a project
      operationId: rename_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameProjectRequest'
        required: true
      responses:
        '200':
          description: Project renamed successfully
...
components:
  schemas:
...
    CreateProjectRequest:
      type: object
      required:
      - project-name
      properties:
        project-id:
          type: string
          format: uuid
          description: |-
            Request a specific project ID - optional.
            If not provided, a new project ID will be generated (recommended).
          nullable: true
        project-name:
          type: string
          description: Name of the project to create.
...
    GetProjectResponse:
      type: object
      required:
      - project_id
      - project_name
      properties:
        project_id:
          type: string
          format: uuid
          description: ID of the project.
        project_name:
          type: string
          description: Name of the project
...
    ListProjectsResponse:
      type: object
      required:
      - projects
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/GetProjectResponse'
          description: List of projects
...